pyproject.toml
[tool.poetry]
name = "photoshare"
version = "1"
description = "backend part"
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.40"
psycopg2-binary = "^2.9.10"
pydantic = "^2.11.1"
alembic = "^1.15.2"
uvicorn = "^0.34.0"
python-dotenv = "^1.1.0"
passlib = "^1.7.4"
python-jose = "^3.4.0"
pydantic-settings = "^2.8.1"
cloudinary = "^1.43.0"
fastapi-mail = "^1.4.2"
slowapi = "^0.1.9"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
